+++ CHANGES.rst
@@ -4,6 +4,15 @@
 New Features
 ^^^^^^^^^^^^
 
+- Allow initialization of ``ImageFileCollection`` from a list of files with no
+  location set. [#374, #661, #680]
+
+- Allow identification of FITS files in ``ImageFileCollection`` based on content
+  of the files instead of file name extension. [#620, #680]
+
+- Add option to use regular expression matching when filtering items in
+  ``ImageFileCollection``. [#480, #595, #682]
+
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
@@ -21,16 +30,28 @@ Other Changes and Additions
 
 - Added warning that ``transform_image`` does not apply the transformation to the WCS [#684]
 
+- Removed support for initializing ``ImageFileCollection`` from a table instead
+  of files. [#680]
+
+- More consistent typing of ``ImageFileCollection.summary`` when the collection
+  is empty. [#601, #680]
 
 Bug Fixes
 ^^^^^^^^^
+
 - Function ``median_combine`` now correctly calculates the uncertainty for
   masked ``CCDData``. [#608]
 
 - Function ``combine`` avoids keeping files open unnecessarily. [#629, #630]
 
+- Function ``combine`` more accurately estimates memory use
+  when deciding how to chunk files. [#638, #642]
+
 - Raise ``ValueError`` error in ``subtract_dark`` for when the errors have
   different shapes [#674, #677]
+
+- Fix problem with column dtypes when initializing ``ImageFileCollection`` from
+  a list of file names. [#662, #680]
 
 1.3.0 (2017-11-1)
 -----------------